Changed portal color to blue and better matched the particle sizes to the portal generator.
