--- sg1/Assets/Cool Visual Effects - Part 1.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8c14b39fcfe4a0b49a0bbdf4af29cbf3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
