--- sg1/Assets/Cool Visual Effects - Part 1/Effects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 78388a007cb89a242a1367f32b18c079
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
