--- sg1/Assets/Cool Visual Effects - Part 1/Effects/Magic Ball.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e2092a594e0ea6f40a7b99fd4251b58d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
