--- sg1/Assets/Cool Visual Effects - Part 1/Effects/Portals.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 89af5cf054c19af49b7d22d98480d51d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
