--- sg1/Assets/Cool Visual Effects - Part 1/Effects/Portals/Portal1.vfx.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: 0cc83dfeaae0cac4caec995a84d8efa3
-VisualEffectImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 176571
-  packageName: Cool Visual Effects - Part 1 - URP Support
-  packageVersion: 2.0
-  assetPath: Assets/Cool Visual Effects - Part 1/Effects/Portals/Portal1.vfx
-  uploadId: 388834
