--- sg1/Assets/Cool Visual Effects - Part 1/Effects/Portals/Portal2.vfx.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: 731ed5960ed955243910de1f66a2c19e
-VisualEffectImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 176571
-  packageName: Cool Visual Effects - Part 1 - URP Support
-  packageVersion: 2.0
-  assetPath: Assets/Cool Visual Effects - Part 1/Effects/Portals/Portal2.vfx
-  uploadId: 388834
