--- sg1/Assets/Cool Visual Effects - Part 1/Effects/Starfield.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 75cb8a4ef5519d547a09c0fb8e1017bf
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
