--- sg1/Assets/Cool Visual Effects - Part 1/Effects/Starfield/Starfield.vfx.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: cd9df07b08ded8b4fa9dde45a583bd5a
-VisualEffectImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 176571
-  packageName: Cool Visual Effects - Part 1 - URP Support
-  packageVersion: 2.0
-  assetPath: Assets/Cool Visual Effects - Part 1/Effects/Starfield/Starfield.vfx
-  uploadId: 388834
