--- sg1/Assets/Cool Visual Effects - Part 1/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5bdd0573642b22341aca716fa448ae96
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
