--- sg1/Assets/Cool Visual Effects - Part 1/Materials/Cold Night by AllSkyFree.mat.meta
@@ -1,15 +1,0 @@
-fileFormatVersion: 2
-guid: 71e1b44e9c7459940875167dc308d849
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 176571
-  packageName: Cool Visual Effects - Part 1 - URP Support
-  packageVersion: 2.0
-  assetPath: Assets/Cool Visual Effects - Part 1/Materials/Cold Night by AllSkyFree.mat
-  uploadId: 388834
