--- sg1/Assets/Cool Visual Effects - Part 1/Materials/Ground.mat.meta
@@ -1,15 +1,0 @@
-fileFormatVersion: 2
-guid: 4b12698b308bc8a468677359b503acb4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 176571
-  packageName: Cool Visual Effects - Part 1 - URP Support
-  packageVersion: 2.0
-  assetPath: Assets/Cool Visual Effects - Part 1/Materials/Ground.mat
-  uploadId: 388834
