--- sg1/Assets/Cool Visual Effects - Part 1/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d954fdcbfd13c4e41b785f3c38af5a2f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
