--- sg1/Assets/Cool Visual Effects - Part 1/Scripts/ConstantRotation.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ConstantRotation : MonoBehaviour
-{
-    public float rotationSpeed = 10.0f;
-    private void Update()
-    {
-        transform.Rotate(Vector3.up * rotationSpeed * Time.deltaTime);
-    }
-}
