--- sg1/Assets/Cool Visual Effects - Part 1/Textures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bc82582740953424599e10f2b8328c72
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
