+++ sg1/Assets/Scenes/Prefabs/Portal End/ConstantRotation.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ConstantRotation : MonoBehaviour
+{
+    public float rotationSpeed = 10.0f;
+    private void Update()
+    {
+        transform.Rotate(Vector3.up * rotationSpeed * Time.deltaTime);
+    }
+}
