+++ sg1/Assets/Scripts/Portal End/PortalEffect.vfx.meta
@@ -1,0 +1,14 @@
+fileFormatVersion: 2
+guid: 851c87b601a57e443b19b37b6ae92621
+VisualEffectImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+AssetOrigin:
+  serializedVersion: 1
+  productId: 176571
+  packageName: Cool Visual Effects - Part 1 - URP Support
+  packageVersion: 2.0
+  assetPath: Assets/Cool Visual Effects - Part 1/Effects/Magic Ball/Magic Ball.vfx
+  uploadId: 388834
